+++ src/payment-flows/native/qrcode.js
@@ -10,6 +10,8 @@ import { getLogger, getStorageID } from '../../lib';
 import { FPTI_STATE, FPTI_TRANSITION, TARGET_ELEMENT, QRCODE_STATE, FPTI_CUSTOM_KEY } from '../../constants';
 import type { ButtonProps, ServiceData, Config, Components } from '../../button/props';
 import { type OnShippingChangeData } from '../../props/onShippingChange';
+import type { Payment } from '../types';
+import { checkout } from '../checkout';
 
 import { getNativeUrl } from './url';
 import { connectNative } from './socket';
@@ -77,7 +79,7 @@ type NativeQRCodeOptions = {|
     serviceData : ServiceData,
     config : Config,
     components : Components,
-    fundingSource : $Values<typeof FUNDING>,
+    payment : Payment,
     sessionUID : string,
     clean : CleanupType,
     callbacks : {|
@@ -115,8 +117,7 @@ type NativeQRCodeOptions = {|
             buttonSessionID : string
         |}>,
         onClose : () => ZalgoPromise<void>,
-        onDestroy : () => ZalgoPromise<void>,
-        onQrEscapePath : (selectedFundingSource : $Values<typeof FUNDING>) => ZalgoPromise<void>
+        onDestroy : () => ZalgoPromise<void>
     |}
 |};
 
@@ -125,10 +126,11 @@ type NativeQRCode = {|
     start : () => ZalgoPromise<void>
 |};
 
-export function initNativeQRCode({ props, serviceData, config, components, fundingSource, clean, callbacks, sessionUID } : NativeQRCodeOptions) : NativeQRCode {
+export function initNativeQRCode({ props, serviceData, config, components, payment, clean, callbacks, sessionUID } : NativeQRCodeOptions) : NativeQRCode {
     const { createOrder, onClick } = props;
     const { QRCode } = components;
-    const { onInit, onApprove, onCancel, onError, onFallback, onClose, onDestroy, onShippingChange, onQrEscapePath } = callbacks;
+    const { fundingSource } = payment;
+    const { onInit, onApprove, onCancel, onError, onFallback, onClose, onDestroy, onShippingChange } = callbacks;
 
     const qrCodeRenderTarget = window.xprops.getParent();
     const pageUrl = window.xprops.getPageUrl();
@@ -161,7 +163,15 @@ export function initNativeQRCode({ props, serviceData, config, components, fundi
                     [FPTI_KEY.TRANSITION]:  `${ FPTI_TRANSITION.QR_PROCESS_PAY_WITH }_${ selectedFundingSource }`
                 }).flush();
 
-                return onQrEscapePath(selectedFundingSource);
+                return ZalgoPromise.try(() => {
+                    const paymentInfo = { ...payment, fundingSource: selectedFundingSource };
+                    const instance = checkout.init({ props, components, payment: paymentInfo, config, serviceData });
+                    clean.register(() => instance.close());
+                    
+                    return instance.start().then(() => {
+                        return ZalgoPromise.resolve();
+                    });
+                });
             };
 
             const validatePromise = ZalgoPromise.try(() => {
